visionOS Media Controls: Controls briefly flash to a different size when you release a resize https://bugs.webkit.org/show_bug.cgi?id=260097 rdar://113771996
Reviewed by NOBODY (OOPS!).

The media controls flashing is just a symptom of the actual bug, which is that when releasing a
resize, the snapshot gets removed too early. This makes it so that after releasing the resize
handle, the video content goes back to its original size briefly before getting back to the final size.

When releasing the resize handle, the live resize ends (`_endLiveResize`). A snapshot view is then
created and shown, to give the web content time to update. Then, theoretically, in the `doAfterNextPresentationUpdate`
completion block, the web content will have been fully laid out and resized from the transaction commit,
and then the snapshot view can be removed.

Howevever, it turns out that `doAfterNextPresentationUpdate` doesn't necessarily guarantee that
everything is laid out. The presentation update causes a transaction commit to happen, and the commit
sets the bounds of the video, which ends up calling `-[CALayer setNeedsLayout]`. This invalidates the
current layout, but does not actually force a layout. Instead, the layout happens at the next
CoreAnimation commit.

What ends up happening is that the presentation update does call `setNeedsLayout` before the completion
block is invoked, but the actual layout of the video, which happens in `-[WebAVPlayerLayer layoutSublayers]`,
does not happen until some time afterwards.

Now, even if `-[WebAVPlayerLayer layoutSublayers]` were too happen before the completion block is invoked,
the bug would still happen. This is because `layoutSublayers` ends up calling `-[WebAVPlayerLayer resolveBounds]`
after a delay, and `resolveBounds` is the method that does the actual resizing.

To illustrate the sequence, this is what is currently happening:
1. The resize ends, which triggers `endLiveResize`
2. A snapshot view is created and placed on top of the web content
3. A presentation update is triggered
4. The presentation update forces a transaction commit to occur
5. The transaction commit updates the bounds of the web content
6. This causes `-[CALayer setNeedsLayout]` to be called
7. After the transaction commit, the presentation update's completion block is invoked
8. The snapshot gets removed. At this point, the video layer's size still has yet to update
9. After some time, `-[WebAVPlayerLayer layoutSublayers]` gets called
10. After even more time, `-[WebAVPlayerLayer resolveBounds]`
11. The video layer's size is finally updated

The solution to this is to ensure steps 9, 10, 11 have to happen strictly before step 8.

To fix, create a completion handler that only gets invoked once the video layer's size has actually been
updated in `-[WebAVPlayerLayer resolveBounds]`, and then don't remove the snapshot until the completion
handler is invoked.

* Source/WebCore/platform/cocoa/VideoFullscreenModel.h:
(WebCore::VideoFullscreenModel::didResolvePlayerLayerBounds):
* Source/WebCore/platform/cocoa/WebAVPlayerLayer.mm:
(-[WebAVPlayerLayer resolveBounds]):
* Source/WebKit/UIProcess/API/ios/WKWebViewIOS.mm:
(-[WKWebView _doAfterVideoFullscreenPlayerLayerBoundsHaveChanged:]):
(-[WKWebView _endLiveResize]):
* Source/WebKit/UIProcess/Cocoa/VideoFullscreenManagerProxy.h:
* Source/WebKit/UIProcess/Cocoa/VideoFullscreenManagerProxy.mm:
(WebKit::VideoFullscreenModelContext::didResolvePlayerLayerBounds):
(WebKit::VideoFullscreenManagerProxy::callAfterPlayerLayerBoundsHaveChanged):
(WebKit::VideoFullscreenManagerProxy::didResolvePlayerLayerBounds):
* Source/WebKit/UIProcess/Cocoa/WebPageProxyCocoa.mm:
(WebKit::WebPageProxy::callAfterVideoFullscreenPlayerLayerBoundsHaveChanged):
* Source/WebKit/UIProcess/WebPageProxy.h: