[GTK] Test fast/picture/viewport-resize.html only fails under Xvfb https://bugs.webkit.org/show_bug.cgi?id=168164
Unreviewed test gardening.

Previously, this test tried to resize the window to 1200x600 but
the biggest resulution we support when running under Xvfb is 1024x768.

The exact window width is not so important for this test.
It just needs to be wider than 800 pixels.

* LayoutTests/fast/picture/resources/resize-test.js:
(standardResizeTest):
* LayoutTests/fast/picture/viewport-resize-expected.txt:
* LayoutTests/platform/gtk/TestExpectations:

Canonical link: https://commits.webkit.org/266906@main