Buttons on the Date/Time picker are too low on visionOS. https://bugs.webkit.org/show_bug.cgi?id=260316 rdar://113725970
Reviewed by Aditya Keerthi.

We had a hard coded size for the accessory view for the buttons
at the bottom of the date/time picker, and since visionOS has
a different layout, it did not look good on the platform. Use
sizeThatFits to have a good size on all platforms and remove
the constant as well as some others that are no longer used.

* Source/WebKit/UIProcess/ios/forms/WKDateTimeInputControl.mm:
(-[WKDateTimePicker initWithView:datePickerMode:]):

Canonical link: https://commits.webkit.org/266973@main