CSS Custom Highlights sometimes does not repaint when live ranges are changed https://bugs.webkit.org/show_bug.cgi?id=259897 rdar://113518997
Reviewed by NOBODY (OOPS!).

[In-Progress Patch]
A page is not aware when a live range is changed, so we never trigger a repaint when live ranges are changed.

The main idea is to repaint the old range, then repaint the new range, similar to Selection which is working as expected.
However, Selection is singular with one range while custom highlights and ranges are a "many to many" relationship.
A range can belong to multiple highlights and a highlight can have multiple ranges.

In this tentative solution, we manually call updateHighlightPositions() whenever a ranges' endpoints are set, which then calls repaint on the old range,
set the range to its new positions, and calls repaint on the new range.
If a range's start and end is set, we will repaint four times (oldSetStartRange, newSetStartRange, oldSetEndRange, newSetEndRange).
For this case, it is better to wait until both points are set to do call repaint (oldRange, newRange).
Unfortunately, since one can set the start only or end only, this "problem" remains unresolved.

A better, more correct solution to the tentative solution is to:
1. Check if the range that is changing belongs to a highlight, since ranges are not highlight-specific.
2. If range belongs to highlight, remember the old range and the new range after it is done being set.
3. Then, signal to the page that there are dirty ranges and we should repaint.

This better, more correct solution is tricky to implement as ranges are set in Range (DOM based), which has no information or knowledge about highlights.
In addition, Range should NOT have information about highlights because a range can belong/be associated with anything, not only highlights.

updateHighlightPositions() belongs in Document (DOM based), and Document does store information about highlights and Page.
The ranges associated with highlights are stored as HighlightRangeData (DOM based) which contains an AbstractRange (DOM based).
- AbstractRange can be static or live ranges.

In the Selection case, SelectionRangeData is a HighlightData.
HighlightData (Render based) is where we would add the logic for repainting old and new ranges.
Given this, a HighlightRangeData should inherit from HighlightData so ranges of highlights can repaint.

Somehow, we should be able to communicate that a highlight range is dirty to the page, which should call HighlightData::repaint()?

* Source/WebCore/dom/Document.cpp:
(WebCore::repaintingRange):
(WebCore::Document::updateHighlightPositions):
* Source/WebCore/dom/Range.cpp:
(WebCore::Range::setStart):
(WebCore::Range::setEnd):