os_log_set_hook return value is improperly ignored https://bugs.webkit.org/show_bug.cgi?id=260303 <rdar://112032774>
Reviewed by Per Arne Vollan.

In Bug 253342 we adopted `os_log_set_hook` so that we could relay logging information
from the WebContent process to the Network process to add to the logging system.

The documentation for this method says:

 * os_log_set_hook returns the previously registered hook or NULL
 * when this function is first called. As a client, if you register a hook and the
 * return value is not NULL then it is your responsibility to chain the
 * previous hook into the new hook so that both are invoked.

Any previously registered hook is unlikely to be able to interact with `logd`, since
we are sandboxed, but perhaps there are other side-effects we want to retain.

To comply with the API design, we should invoke any preregistered hooks before calling
our own logic.

* Source/WebKit/WebProcess/cocoa/WebProcessCocoa.mm:
(WebKit::registerLogHook):

Canonical link: https://commits.webkit.org/267007@main